% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo_ref_loc-data.R
\docType{data}
\name{argo_ref_loc2}
\alias{argo_ref_loc2}
\title{Argo profile dataset around Reference Location 2 at \eqn{(L, l) = (0, 175W)}}
\format{
A list containing numeric vectors of \code{TemperatureResiduals},
\code{SalinityResiduals}, \code{Pressure}, \code{Latitude}, \code{Longitude}, \code{ProfileNumber},
and \code{DistanceKmFromRefLoc}.
}
\source{
\url{https://github.com/dyarger/argofda}
}
\usage{
data(argo_ref_loc2)
}
\description{
A bivariate dataset of temperature and salinity residuals after removing
a space-time functional mean. Residuals are processed and obtained from
raw Argo data by Yarger, D., Stoev, S., & Hsing, T. (2022).
}
\examples{
data(argo_ref_loc2)
\donttest{
# heat map of correlation matrix, linked to scatterplots
head(argo_ref_loc2$TemperatureResiduals)}
}
\references{
Yarger, D., Stoev, S., & Hsing, T. (2022). A functional-data approach to the Argo data. \emph{The Annals of Applied Statistics, 16(1)}, 216-246.
}
\author{
Mary Lai Salvana \email{yourlainess@gmail.com}
}
\keyword{datasets}
