% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_auxilliary.R
\name{bsplineBasis}
\alias{bsplineBasis}
\title{Compute the value of the bspline bases}
\usage{
bsplineBasis(
  x,
  degree,
  innerknots,
  lowknot = min(x, innerknots),
  highknot = max(x, innerknots)
)
}
\arguments{
\item{x}{A matrix of basis function values.}

\item{degree}{A number indicating the degrees.}

\item{innerknots}{A vector of knot locations.}

\item{lowknot}{The first knot location.}

\item{highknot}{The last knot location.}
}
\value{
A matrix of basis function values.
}
\description{
Compute the value of the bspline bases
}
