% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_auxilliary.R
\name{compute_emp_cov}
\alias{compute_emp_cov}
\title{Compute the kernel smooth empirical cross-covariance values}
\usage{
compute_emp_cov(location, variable1_residuals, variable2_residuals,
bandwidth_horizontal, bandwidth_vertical, radius)
}
\arguments{
\item{location}{An \eqn{n \times 3} matrix of coordinates.}

\item{variable1_residuals}{A numeric vector of variable 1 residuals.}

\item{variable2_residuals}{A numeric vector of variable 2 residuals.}

\item{bandwidth_horizontal}{A numeric constant parameter for the bandwidth
in the horizontal direction.}

\item{bandwidth_vertical}{A numeric constant parameter for the bandwidth
in the vertical direction.}

\item{radius}{A numeric constant indicating the radius of the sphere.}
}
\value{
A matrix of dimension \eqn{2 n \times 2 n}.
}
\description{
\code{compute_emp_cov} computes the kernel smoothed empirical cross-covariance,
\eqn{\hat{C}_{ij}(\mathbf{s}_1, \mathbf{s}_2)} between any two locations
\eqn{\mathbf{s}_1} and \eqn{\mathbf{s}_2}, which is required by the
weighted least squares estimation routine.
The formula used to compute \eqn{\hat{C}_{ij}(\mathbf{s}_1, \mathbf{s}_2)}
was proposed by Kleiber, W., & Nychka, D. (2012) and has the following form:
\deqn{\hat{C}_{ij}(\mathbf{s}_1, \mathbf{s}_2) = \frac{\sum_{\tilde{\mathbf{s}} \in \tilde{\mathcal{D}}} K_{\lambda_{h}, \lambda_{v}}(\mathbf{s}_1, \tilde{\mathbf{s}})^{1/2} K_{\lambda_{h}, \lambda_{v}}(\mathbf{s}_2, \tilde{\mathbf{s}})^{1/2} Z_i (\tilde{\mathbf{s}}) Z_j (\tilde{\mathbf{s}})}{ \{\sum_{\tilde{\mathbf{s}} \in \tilde{\mathcal{D}}} K_{\lambda_{h}, \lambda_{v}}(\mathbf{s}_1, \tilde{\mathbf{s}})\}^{1/2} \{\sum_{\tilde{\mathbf{s}} \in \tilde{\mathcal{D}}} K_{\lambda_{h}, \lambda_{v}}(\mathbf{s}_2, \tilde{\mathbf{s}})\}^{1/2}},}
where \eqn{K_{\lambda_{h}, \lambda_{v}}} is a nonnegative kernel function
with horizontal and vertical bandwidths, \eqn{\lambda_{h}} and
\eqn{\lambda_{v}}, respectively, which control the degree of smoothing.
}
\examples{

data(argo_ref_loc1)

loc3d <- cbind(argo_ref_loc1$Longitude, argo_ref_loc1$Latitude, argo_ref_loc1$Pressure)

earthRadiusKm = 6371

\dontrun{emp_cov <- compute_emp_cov(location = loc3d,
variable1_residuals = argo_ref_loc1$TemperatureResiduals,
variable2_residuals = argo_ref_loc1$SalinityResiduals,
bandwidth_horizontal = 0.009, bandwidth_vertical = 0.03,
radius = earthRadiusKm)}


}
\references{
Kleiber, W., & Nychka, D. (2012). Nonstationary modeling for multivariate spatial processes. \emph{Journal of Multivariate Analysis, 112}, 76-91.
}
\author{
Mary Lai Salvana \email{yourlainess@gmail.com}
}
