% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_auxilliary.R
\name{est_bi_differential}
\alias{est_bi_differential}
\title{Maximum Likelihood Estimation of the parameters of the bivariate differential operator cross-covariance function}
\usage{
est_bi_differential(residuals, location, init_beta, init_scale_horizontal, 
init_scale_vertical, init_scale_horizontal_fix, init_scale_vertical_fix,
init_a1, init_b1, init_c1, init_d1, init_a1_fix, init_b1_fix, 
init_a2, init_b2, init_c2, init_d2, init_a2_fix, init_b2_fix, 
init_beta_fix, radius, basis1, nb1, basis2, nb2, splines_degree,
knots1, knots2, MAXIT, RERUNS, STEPMAX)
}
\arguments{
\item{residuals}{A vector of residuals}

\item{location}{An nx3 matrix of coordinates.}

\item{init_beta}{A number for colocated correlation parameter.}

\item{init_scale_horizontal}{A number for the horizontal scale parameter.}

\item{init_scale_vertical}{A number for the vertical scale parameter.}

\item{init_scale_horizontal_fix}{An indicator whether horizontal scale parameter should be estimated.}

\item{init_scale_vertical_fix}{An indicator whether vertical scale parameter should be estimated.}

\item{init_a1}{A number for the anisotropy parameter in Latitude associated with variable 1.}

\item{init_b1}{A number for the anisotropy parameter in longitude associated with variable 1.}

\item{init_c1}{A number for vector for the nonstationarity parameter in depth associated with variable 1.}

\item{init_d1}{A number for the variance parameter of the fully isotropic associated with variable 1.}

\item{init_a1_fix}{An indicator whether a1 parameter should be estimated.}

\item{init_b1_fix}{An indicator whether b1 parameter should be estimated.}

\item{init_a2}{A number for the anisotropy parameter in Latitude associated with variable 2.}

\item{init_b2}{A number for the anisotropy parameter in longitude associated with variable 2.}

\item{init_c2}{A number for vector for the nonstationarity parameter in depth associated with variable 2.}

\item{init_d2}{A number for the variance parameter of the fully isotropic associated with variable 2.}

\item{init_a2_fix}{An indicator whether a1 parameter should be estimated.}

\item{init_b2_fix}{An indicator whether b1 parameter should be estimated.}

\item{init_beta_fix}{An indicator whether colocated correlation parameter should be estimated.}

\item{radius}{A number for the radius of the sphere.}

\item{basis1}{A matrix of basis function values for variable 1.}

\item{nb1}{A number indicating the number of bases for variable 1.}

\item{basis2}{A matrix of basis function values for variable 2.}

\item{nb2}{A number indicating the number of bases for variable 2.}

\item{splines_degree}{A number indicating the degree of the splines.}

\item{knots1}{A vector of knot locations for variable 1.}

\item{knots2}{A vector of knot locations for variable 2.}

\item{MAXIT}{A number indicating the maximum number of iterations for optim.}

\item{RERUNS}{A number indicating the number of times optim is re-run from previous MLE.}

\item{STEPMAX}{A number indicating the stepmax of nlm.}
}
\value{
A vector of estimated parameter values.
}
\description{
\code{est_bi_differential} performs maximum likelihood estimation using a
Newton-type algorithm to carry out a minimization of the negative of the
Gaussian log-likelihood function:
\deqn{l(\boldsymbol{\theta}) = -\frac{q n}{2} \log (2 \pi) - \frac{1}{2} \log |\boldsymbol{\Sigma}(\boldsymbol{\theta})| - \frac{1}{2} \mathbf{Z}^{\top} \boldsymbol{\Sigma}(\boldsymbol{\theta})^{-1} \mathbf{Z},}
Here \eqn{q} is the number of parameters and \eqn{\boldsymbol{\Sigma}}.
}
