% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_auxilliary.R
\name{est_bi_differential_mle}
\alias{est_bi_differential_mle}
\title{Maximum Likelihood Estimation of the parameters of the bivariate differential operator cross-covariance function}
\usage{
est_bi_differential_mle(residuals, location, init_beta, init_scale_horizontal,
init_scale_vertical, init_a1, init_b1, init_c1, init_d1,
init_a2, init_b2, init_c2, init_d2,
beta_fix, scale_horizontal_fix, scale_vertical_fix,
a1_fix, b1_fix, c1_fix, d1_fix, a2_fix, b2_fix, c2_fix, d2_fix,
beta_scaling, horizontal_scale_scaling, vertical_scale_scaling,
a1_scaling, b1_scaling, c1_scaling, d1_scaling,
a2_scaling, b2_scaling, c2_scaling, d2_scaling,
radius_of_sphere, splines_degree, inner_knots1, inner_knots2, iterlim, stepmax, hessian)
}
\arguments{
\item{residuals}{A \eqn{2 n} vector of residuals with
variable 1 and 2 residuals appended next to each other.}

\item{location}{An \eqn{n \times 3} matrix of coordinates.}

\item{init_beta}{A numeric constant indicating the initial value for the
colocated correlation parameter.}

\item{init_scale_horizontal}{A numeric constant indicating the initial value for the
horizontal scale parameter.}

\item{init_scale_vertical}{A numeric constant indicating the initial value for the
vertical scale parameter.}

\item{init_a1}{A numeric constant indicating the initial value for the
anisotropy in latitude parameter associated with variable 1.}

\item{init_b1}{A numeric constant indicating the initial value for the
anisotropy in longitude parameter associated with variable 1.}

\item{init_c1}{A numeric vector indicating the initial value for the
splines coefficients for the nonstationary with depth parameter c1
associated with variable 1.}

\item{init_d1}{A numeric constant indicating the initial value for the
variance parameter from the fully isotropic component associated with variable 1.}

\item{init_a2}{A numeric constant indicating the initial value for the
anisotropy in latitude parameter associated with variable 2.}

\item{init_b2}{A numeric constant indicating the initial value for the
anisotropy in longitude parameter associated with variable 2.}

\item{init_c2}{A numeric vector indicating the initial value for the
splines coefficients for the nonstationary with depth parameter c2
associated with variable 2.}

\item{init_d2}{A numeric constant indicating the initial value for the
variance parameter from the fully isotropic component associated with variable 2.}

\item{beta_fix}{An indicator that takes in the values \code{TRUE} or \code{FALSE} whether the
colocated correlation parameter should not be estimated.}

\item{scale_horizontal_fix}{An indicator that takes in the values \code{TRUE} or \code{FALSE} whether the
horizontal scale parameter should not be estimated.}

\item{scale_vertical_fix}{An indicator that takes in the values \code{TRUE} or \code{FALSE} whether the
vertical scale parameter should not be estimated.}

\item{a1_fix}{If \code{TRUE}, the a1 parameter is not be estimated.}

\item{b1_fix}{If \code{TRUE}, the b1 parameter is not be estimated.}

\item{c1_fix}{If \code{TRUE}, the c1 parameters are not be estimated.}

\item{d1_fix}{If \code{TRUE}, the d1 parameter is not be estimated.}

\item{a2_fix}{If \code{TRUE}, the a2 parameter is not be estimated.}

\item{b2_fix}{If \code{TRUE}, the b2 parameter is not be estimated.}

\item{c2_fix}{If \code{TRUE}, the c2 parameters are not be estimated.}

\item{d2_fix}{If \code{TRUE}, the d2 parameter is not be estimated.}

\item{beta_scaling}{A numeric constant indicating the scaling applied to the
colocated correlation parameter}

\item{horizontal_scale_scaling}{A numeric constant indicating the scaling
applied to the horizontal scale parameter.}

\item{vertical_scale_scaling}{A numeric constant indicating the scaling
applied to the vertical scale parameter.}

\item{a1_scaling}{A numeric constant indicating the scaling applied to the
anisotropy in latitude parameter associated with variable 1.}

\item{b1_scaling}{A numeric constant indicating the scaling applied to the
anisotropy in longitude parameter associated with variable 1.}

\item{c1_scaling}{A numeric constant indicating the scaling
applied to the splines coefficients for the nonstationary with depth
parameter c1 associated with variable 1.}

\item{d1_scaling}{A numeric constant indicating the scaling applied to the
variance parameter from the fully isotropic component associated with variable 1.}

\item{a2_scaling}{A numeric constant indicating the scaling applied to the
anisotropy in latitude parameter associated with variable 2.}

\item{b2_scaling}{A numeric constant indicating the scaling applied to the
anisotropy in longitude parameter associated with variable 2.}

\item{c2_scaling}{A numeric constant indicating the scaling
applied to the splines coefficients for the nonstationary with depth
parameter c2 associated with variable 2.}

\item{d2_scaling}{A numeric constant indicating the scaling applied to the
variance parameter from the fully isotropic component associated with variable 2.}

\item{radius_of_sphere}{A numeric constant indicating the radius of the sphere.}

\item{splines_degree}{A number indicating the degree of the splines.}

\item{inner_knots1}{A vector of inner knot locations for variable 1.}

\item{inner_knots2}{A vector of inner knot locations for variable 2.}

\item{iterlim}{A number indicating the maximum number of iterations for \code{nlm}.}

\item{stepmax}{A number indicating the stepmax of nlm.}

\item{hessian}{If \code{TRUE}, the hessian is not returned.}
}
\value{
A list containing the values of the convergence code,
log likelihood function,
estimated parameters and their corresponding standard deviations,
splines degree, and pre-specified knots.
}
\description{
\code{est_bi_differential_mle} performs maximum likelihood estimation using a
Newton-type algorithm to carry out a minimization of the negative of the
Gaussian log-likelihood function:
\deqn{l(\boldsymbol{\theta}) = -\frac{q n}{2} \log (2 \pi) - \frac{1}{2} \log |\boldsymbol{\Sigma}(\boldsymbol{\theta})| - \frac{1}{2} \mathbf{Z}^{\top} \boldsymbol{\Sigma}(\boldsymbol{\theta})^{-1} \mathbf{Z},}
Here \eqn{\boldsymbol{\theta}} is a vector parameters,
\eqn{q} is the number of parameters,
\eqn{\boldsymbol{\Sigma}(\boldsymbol{\theta})}
is the \eqn{q n \times q n} cross-covariance matrix for \eqn{\mathbf{Z}}
with determinant \eqn{|\boldsymbol{\Sigma}(\boldsymbol{\theta})|}, and
\eqn{\mathbf{Z} = \{\mathbf{Z}(\mathbf{s}_1)^{\top}, \ldots, \mathbf{Z}(\mathbf{s}_n)^{\top} \}^{\top}}
is the vector of residuals such that \eqn{\mathbf{Z}(\mathbf{s}) = \{ Z_1 (\mathbf{s}), Z_2 (\mathbf{s})\}^{\top}}.
}
\examples{

library(dplyr)

x <- seq(0, 1, length.out = 5)
y <- seq(0, 1, length.out = 5)

loc2d <- expand.grid(x, y) \%>\% as.matrix()
depth <- seq(0, 1, length.out = 5)
loc3d <- cbind(rep(loc2d[, 1], each = length(depth)), rep(loc2d[, 2], each = length(depth)), depth)

earthRadiusKm = 6371

BETA = 0.5
SCALE_HORIZONTAL = 0.03
SCALE_VERTICAL = 0.3
A1 = A2 = 0.00001
B1 = B2 = 0.00001
C1 = sin((loc3d[, 3] + 0.1) * pi / 0.5)
C2 = cos((loc3d[, 3] + 0.1) * pi / 0.5)
D1 = D2 = 0

cov_mat <- cov_bi_differential(location = loc3d, beta = BETA,
                               scale_horizontal = SCALE_HORIZONTAL,
                               scale_vertical = SCALE_VERTICAL,
                               a1 = A1, b1 = B1, c1 = C1, d1 = D1,
                               a2 = A2, b2 = B2, c2 = C2, d2 = D2,
                               radius_of_sphere = earthRadiusKm)


library(MASS)

set.seed(1235)
Z <- mvrnorm(1, mu = rep(0, ncol(cov_mat)), Sigma = cov_mat)
Z1 <- Z[1:nrow(loc3d)]
Z2 <- Z[nrow(loc3d) + 1:nrow(loc3d)]

INNER_KNOTS1 <- c(0.1, 0.5, 0.9)
INNER_KNOTS2 <- c(0.1, 0.5, 0.9)

INIT_BETA = 0.6
INIT_SCALE_HORIZONTAL = log(0.02)
INIT_SCALE_VERTICAL = log(0.4)
INIT_A1 = INIT_A2 = 0.005
INIT_B1 = INIT_B2 = 0.005
INIT_D1 = INIT_D2 = 0

SPLINES_DEGREE = 2

set.seed(1235)
INIT_C1 <- runif(length(INNER_KNOTS1) + SPLINES_DEGREE + 1, -5, 5)

set.seed(1236)
INIT_C2 <- runif(length(INNER_KNOTS2) + SPLINES_DEGREE + 1, -5, 5)

\dontrun{est_params_mle <- est_bi_differential_mle(residuals = Z, location = loc3d, init_beta = INIT_BETA,
                                 init_scale_horizontal = INIT_SCALE_HORIZONTAL,
                                 init_scale_vertical = INIT_SCALE_VERTICAL,
                                 init_a1 = INIT_A1, init_b1 = INIT_B1,
                                 init_c1 = INIT_C1, init_d1 = INIT_D1,
                                 init_a2 = INIT_A2, init_b2 = INIT_B2,
                                 init_c2 = INIT_C2, init_d2 = INIT_D2,
                                 d1_fix = TRUE, d2_fix = TRUE,
                                 a1_scaling = 1e-3, b1_scaling = 1e-3,
                                 a2_scaling = 1e-3, b2_scaling = 1e-3,
                                 radius_of_sphere = earthRadiusKm,
                                 splines_degree = SPLINES_DEGREE,
                                 inner_knots1 = INNER_KNOTS1,
                                 inner_knots2 = INNER_KNOTS2,
                                 iterlim = 1, hessian = T)}


}
\author{
Mary Lai Salvana \email{yourlainess@gmail.com}
}
