% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_auxilliary.R
\name{plot_bi_differential}
\alias{plot_bi_differential}
\title{Plotting the bivariate differential operator cross-covariance function}
\usage{
plot_bi_differential(
  location,
  est_beta,
  est_scale_horizontal,
  est_scale_vertical,
  est_a1,
  est_b1,
  est_c1,
  est_d1 = NULL,
  est_a2,
  est_b2,
  est_c2,
  est_d2 = NULL,
  radius,
  basis1,
  nb1 = ncol(basis1),
  basis2,
  nb2 = ncol(basis2),
  splines_degree = 4
)
}
\arguments{
\item{location}{An nx3 matrix of coordinates with latitude and longitude of the reference location and the pressure coordinate is ordered from surface to bottom.}

\item{est_beta}{A number for colocated correlation parameter.}

\item{est_scale_horizontal}{A number for the horizontal scale parameter.}

\item{est_scale_vertical}{A number for the vertical scale parameter.}

\item{est_a1}{A number for the anisotropy parameter in Latitude associated with variable 1.}

\item{est_b1}{A number for the anisotropy parameter in longitude associated with variable 1.}

\item{est_c1}{A number for vector for the nonstationarity parameter in depth associated with variable 1.}

\item{est_d1}{A number for the variance parameter of the fully isotropic associated with variable 1.}

\item{est_a2}{A number for the anisotropy parameter in Latitude associated with variable 2.}

\item{est_b2}{A number for the anisotropy parameter in longitude associated with variable 2.}

\item{est_c2}{A number for vector for the nonstationarity parameter in depth associated with variable 2.}

\item{est_d2}{A number for the variance parameter of the fully isotropic associated with variable 2.}

\item{radius}{A number for the radius of the sphere.}

\item{basis1}{A matrix of basis function values for variable 1.}

\item{nb1}{A number indicating the number of bases for variable 1.}

\item{basis2}{A matrix of basis function values for variable 2.}

\item{nb2}{A number indicating the number of bases for variable 2.}

\item{splines_degree}{A number indicating the degree of the splines.}
}
\value{
Figure/plots
}
\description{
Plotting the bivariate differential operator cross-covariance function
}
