% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo_auxiliary_functions.R
\name{plotSurfaceMap}
\alias{plotSurfaceMap}
\title{Plot surface maps}
\usage{
plotSurfaceMap(
  variable = "temp",
  year = NA,
  profJulFracofYearAggr,
  profYearAggr,
  profLatAggr,
  profLongAggr,
  profTempAggr,
  profPsalAggr
)
}
\arguments{
\item{variable}{A matrix of basis function values.}

\item{year}{A matrix of basis function values.}

\item{profJulFracofYearAggr}{A matrix of basis function values.}

\item{profYearAggr}{A matrix of basis function values.}

\item{profLatAggr}{A matrix of basis function values.}

\item{profLongAggr}{A matrix of basis function values.}

\item{profTempAggr}{A matrix of basis function values.}

\item{profPsalAggr}{A matrix of basis function values.}
}
\description{
Plot surface maps
}
