% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_auxilliary.R
\name{est_bi_differential_wls}
\alias{est_bi_differential_wls}
\title{Weighted Least Squares Estimation of the parameters of the bivariate differential operator cross-covariance function}
\usage{
est_bi_differential_wls(empirical_values, location, init_beta, init_scale_horizontal,
init_scale_vertical, init_a1, init_b1, init_c1, init_d1,
init_a2, init_b2, init_c2, init_d2,
beta_fix, scale_horizontal_fix, scale_vertical_fix,
a1_fix, b1_fix, c1_fix, d1_fix, a2_fix, b2_fix, c2_fix, d2_fix,
beta_scaling, horizontal_scale_scaling, vertical_scale_scaling,
a1_scaling, b1_scaling, c1_scaling, d1_scaling,
a2_scaling, b2_scaling, c2_scaling, d2_scaling,
radius_of_sphere, splines_degree, inner_knots1, inner_knots2,
w1, w2, w12, iterlim, stepmax, hessian)
}
\arguments{
\item{empirical_values}{A matrix of dimension \eqn{2 n \times 2 n} containing
the empirical marginal and cross-covariance values.}

\item{location}{An \eqn{n \times 3} matrix of coordinates.}

\item{init_beta}{A numeric constant indicating the initial value for the
colocated correlation parameter.}

\item{init_scale_horizontal}{A numeric constant indicating the initial value for the
horizontal scale parameter.}

\item{init_scale_vertical}{A numeric constant indicating the initial value for the
vertical scale parameter.}

\item{init_a1}{A numeric constant indicating the initial value for the
anisotropy in latitude parameter associated with variable 1.}

\item{init_b1}{A numeric constant indicating the initial value for the
anisotropy in longitude parameter associated with variable 1.}

\item{init_c1}{A numeric vector indicating the initial value for the
splines coefficients for the nonstationary with depth parameter c1
associated with variable 1.}

\item{init_d1}{A numeric constant indicating the initial value for the
variance parameter from the fully isotropic component associated with variable 1.}

\item{init_a2}{A numeric constant indicating the initial value for the
anisotropy in latitude parameter associated with variable 2.}

\item{init_b2}{A numeric constant indicating the initial value for the
anisotropy in longitude parameter associated with variable 2.}

\item{init_c2}{A numeric vector indicating the initial value for the
splines coefficients for the nonstationary with depth parameter c2
associated with variable 12.}

\item{init_d2}{A numeric constant indicating the initial value for the
variance parameter from the fully isotropic component associated with variable 2.}

\item{beta_fix}{An indicator that takes in the values \code{TRUE} or \code{FALSE} whether the
colocated correlation parameter should not be estimated.}

\item{scale_horizontal_fix}{An indicator that takes in the values \code{TRUE} or \code{FALSE} whether the
horizontal scale parameter should not be estimated.}

\item{scale_vertical_fix}{An indicator that takes in the values \code{TRUE} or \code{FALSE} whether the
vertical scale parameter should not be estimated.}

\item{a1_fix}{If \code{TRUE}, the a1 parameter is not be estimated.}

\item{b1_fix}{If \code{TRUE}, the b1 parameter is not be estimated.}

\item{c1_fix}{If \code{TRUE}, the c1 parameters are not be estimated.}

\item{d1_fix}{If \code{TRUE}, the d1 parameter is not be estimated.}

\item{a2_fix}{If \code{TRUE}, the a2 parameter is not be estimated.}

\item{b2_fix}{If \code{TRUE}, the b2 parameter is not be estimated.}

\item{c2_fix}{If \code{TRUE}, the c2 parameters are not be estimated.}

\item{d2_fix}{If \code{TRUE}, the d2 parameter is not be estimated.}

\item{beta_scaling}{A numeric constant indicating the scaling applied to the
colocated correlation parameter}

\item{horizontal_scale_scaling}{A numeric constant indicating the scaling
applied to the horizontal scale parameter.}

\item{vertical_scale_scaling}{A numeric constant indicating the scaling
applied to the vertical scale parameter.}

\item{a1_scaling}{A numeric constant indicating the scaling applied to the
anisotropy in latitude parameter associated with variable 1.}

\item{b1_scaling}{A numeric constant indicating the scaling applied to the
anisotropy in longitude parameter associated with variable 1.}

\item{c1_scaling}{A numeric constant indicating the scaling
applied to the splines coefficients for the nonstationary with depth
parameter c1 associated with variable 1.}

\item{d1_scaling}{A numeric constant indicating the scaling applied to the
variance parameter from the fully isotropic component associated with variable 1.}

\item{a2_scaling}{A numeric constant indicating the scaling applied to the
anisotropy in latitude parameter associated with variable 2.}

\item{b2_scaling}{A numeric constant indicating the scaling applied to the
anisotropy in longitude parameter associated with variable 2.}

\item{c2_scaling}{A numeric constant indicating the scaling
applied to the splines coefficients for the nonstationary with depth
parameter c2 associated with variable 2.}

\item{d2_scaling}{A numeric constant indicating the scaling applied to the
variance parameter from the fully isotropic component associated with variable 2.}

\item{radius_of_sphere}{A numeric constant indicating the radius of the sphere.}

\item{splines_degree}{A number indicating the degree of the splines.}

\item{inner_knots1}{A vector of knot locations for variable 1.}

\item{inner_knots2}{A vector of knot locations for variable 2.}

\item{w1}{A number indicating the weight for the marginal covariance
for variable 1 in the weighted least squares objective function.}

\item{w2}{A number indicating the weight for the marginal covariance
for variable 2 in the weighted least squares objective function.}

\item{w12}{A number indicating the weight for the cross-covariance
in the weighted least squares objective function.}

\item{iterlim}{A number indicating the maximum number of iterations for \code{nlm}.}

\item{stepmax}{A number indicating the stepmax of nlm.}

\item{hessian}{If \code{TRUE}, the hessian is not returned.}
}
\value{
A list containing the values of the convergence code,
minimum of the objective function,
estimated parameters and their corresponding standard deviations,
splines degree, and pre-specified knots.
}
\description{
\code{est_bi_differential_wls} performs weighted least squares estimation using a
Newton-type algorithm to carry out a minimization of the objective function:
\deqn{\mathcal{Q}(\boldsymbol{\theta}) = \sum_{i, j = 1}^{q} \sum_{a = 1}^{n} \sum_{b = 1}^{n} w_{ij} \left\{ \hat{C}_{ij}(\mathbf{s}_a, \mathbf{s}_b) - C_{ij}(\mathbf{s}_a, \mathbf{s}_b | \boldsymbol{\theta})\right\}^2,}
where \eqn{\boldsymbol{\theta}} is a vector parameters,
\eqn{q} is the number of parameters,
\eqn{\hat{C}_{ij}(\mathbf{s}_a, \mathbf{s}_b)} is the kernel smoothed
empirical cross-covariance computed from the data, and \eqn{w_{ij}} is the
weight assigned to component \eqn{(i,j)} of this least squares problem.
}
\examples{

data(argo_ref_loc1)

loc3d <- cbind(argo_ref_loc1$Longitude, argo_ref_loc1$Latitude, argo_ref_loc1$Pressure)

earthRadiusKm = 6371

\dontrun{emp_cov <- compute_emp_cov(location = loc3d,
variable1_residuals = argo_ref_loc1$TemperatureResiduals,
variable2_residuals = argo_ref_loc1$SalinityResiduals,
bandwidth_horizontal = 0.009, bandwidth_vertical = 0.03,
radius_of_sphere = earthRadiusKm)}

INNER_KNOTS1 <- c(0.1, 0.5, 0.9)
INNER_KNOTS2 <- c(0.1, 0.5, 0.9)

INIT_BETA = 0.6
INIT_SCALE_HORIZONTAL = log(0.02)
INIT_SCALE_VERTICAL = log(0.4)
INIT_A1 = INIT_A2 = 0.005
INIT_B1 = INIT_B2 = 0.005
INIT_D1 = INIT_D2 = 0

SPLINES_DEGREE = 2

set.seed(1235)
INIT_C1 <- runif(length(INNER_KNOTS1) + SPLINES_DEGREE + 1, -5, 5)

set.seed(1236)
INIT_C2 <- runif(length(INNER_KNOTS2) + SPLINES_DEGREE + 1, -5, 5)

\dontrun{est_params_wls <- est_bi_differential_wls(empirical_values = emp_cov, location = loc3d,
                                 init_beta = INIT_BETA,
                                 init_scale_horizontal = INIT_SCALE_HORIZONTAL,
                                 init_scale_vertical = INIT_SCALE_VERTICAL,
                                 init_a1 = INIT_A1, init_b1 = INIT_B1,
                                 init_c1 = INIT_C1, init_d1 = INIT_D1,
                                 init_a2 = INIT_A2, init_b2 = INIT_B2,
                                 init_c2 = INIT_C2, init_d2 = INIT_D2,
                                 d1_fix = TRUE, d2_fix = TRUE,
                                 a1_scaling = 1e-3, b1_scaling = 1e-3,
                                 a2_scaling = 1e-3, b2_scaling = 1e-3,
                                 radius_of_sphere = earthRadiusKm,
                                 splines_degree = SPLINES_DEGREE,
                                 inner_knots1 = INNER_KNOTS1,
                                 inner_knots2 = INNER_KNOTS2,
                                 w1 = 100, w2 = 50000, w12 = 1000,
                                 iterlim = 1, hessian = FALSE)}

}
\author{
Mary Lai Salvana \email{yourlainess@gmail.com}
}
