% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_auxilliary.R
\name{est_bi_differential_wls}
\alias{est_bi_differential_wls}
\title{Step 1: WLS estimation of the parameters of the bivariate differential operator cross-covariance function}
\usage{
est_bi_differential_wls(
  empirical_values,
  location,
  init_beta,
  init_scale_horizontal,
  init_scale_vertical,
  init_scale_horizontal_fix = F,
  init_scale_vertical_fix = F,
  init_a1,
  init_b1,
  init_c1,
  init_d1 = NULL,
  init_a1_fix = F,
  init_b1_fix = F,
  init_a2,
  init_b2,
  init_c2,
  init_d2 = NULL,
  init_a2_fix = F,
  init_b2_fix = F,
  init_beta_fix = F,
  radius,
  basis1,
  basis2,
  splines_degree = 4,
  knots1,
  knots2
)
}
\arguments{
\item{empirical_values}{A matrix of empirical values}

\item{location}{An nx3 matrix of coordinates.}

\item{init_beta}{A number for colocated correlation parameter.}

\item{init_scale_horizontal}{A number for the horizontal scale parameter.}

\item{init_scale_vertical}{A number for the vertical scale parameter.}

\item{init_scale_horizontal_fix}{An indicator whether horizontal scale parameter should be estimated.}

\item{init_scale_vertical_fix}{An indicator whether vertical scale parameter should be estimated.}

\item{init_a1}{A number for the anisotropy parameter in Latitude associated with variable 1.}

\item{init_b1}{A number for the anisotropy parameter in longitude associated with variable 1.}

\item{init_c1}{A number for vector for the nonstationarity parameter in depth associated with variable 1.}

\item{init_d1}{A number for the variance parameter of the fully isotropic associated with variable 1.}

\item{init_a1_fix}{An indicator whether a1 parameter should be estimated.}

\item{init_b1_fix}{An indicator whether b1 parameter should be estimated.}

\item{init_a2}{A number for the anisotropy parameter in Latitude associated with variable 2.}

\item{init_b2}{A number for the anisotropy parameter in longitude associated with variable 2.}

\item{init_c2}{A number for vector for the nonstationarity parameter in depth associated with variable 2.}

\item{init_d2}{A number for the variance parameter of the fully isotropic associated with variable 2.}

\item{init_a2_fix}{An indicator whether a2 parameter should be estimated.}

\item{init_b2_fix}{An indicator whether b2 parameter should be estimated.}

\item{init_beta_fix}{An indicator whether colocated correlation parameter should be estimated.}

\item{radius}{A number for the radius of the sphere.}

\item{basis1}{A matrix of basis function values for variable 1.}

\item{basis2}{A matrix of basis function values for variable 2.}

\item{splines_degree}{A number indicating the degree of the splines.}

\item{knots1}{A vector of knot locations for variable 1.}

\item{knots2}{A vector of knot locations for variable 2.}
}
\value{
A vector of estimated parameter values.
}
\description{
Step 1: WLS estimation of the parameters of the bivariate differential operator cross-covariance function
}
